- Renamed "Bam" to "Sam" since this is about sequence files (SAM, BAM, and eventually CRAM), not BAM in particular.
