+++ src/main/java/picard/sam/SamErrorMetric/BaseErrorCalculator.java
@@ -22,44 +22,23 @@
  * THE SOFTWARE.
  */
 
-package picard.sam.BamErrorMetric;
+package picard.sam.SamErrorMetric;
 
 import htsjdk.samtools.reference.SamLocusAndReferenceIterator;
 import htsjdk.samtools.util.SamLocusIterator;
 import htsjdk.samtools.util.SequenceUtil;
 
-/**
- * A calculator that estimates the error rate of the bases it observes, assuming that the reference is truth.
- */
-public class SimpleErrorCalculator extends BaseErrorCalculator {
-
-    protected long nMismatchingBases = 0;
+public abstract class BaseErrorCalculator implements BaseCalculator {
+    long totalBases = 0;
 
     /**
-     * The function by which new loci are "shown" to the calculator
+     * the function by which new loci are "shown" to the calculator
      **/
     @Override
     public void addBase(final SamLocusIterator.RecordAndOffset recordAndOffset, final SamLocusAndReferenceIterator.SAMLocusAndReference locusAndRef) {
-        super.addBase(recordAndOffset, locusAndRef);
         final byte readBase = recordAndOffset.getReadBase();
-        if (!SequenceUtil.isNoCall(readBase) && (readBase != locusAndRef.getReferenceBase())) {
-            nMismatchingBases++;
+        if (!SequenceUtil.isNoCall(readBase)) {
+            totalBases++;
         }
-    }
-
-    /**
-     * The suffix that pertains to the implementation of aggregation
-     **/
-    @Override
-    public String getSuffix() {
-        return "error";
-    }
-
-    /**
-     * Returns the metric generated by the observed loci
-     **/
-    @Override
-    public BaseErrorMetric getMetric() {
-        return new BaseErrorMetric("", totalBases, nMismatchingBases);
     }
 }
