+++ src/main/java/picard/sam/SamErrorMetric/Pair.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package picard.sam.BamErrorMetric;
+package picard.sam.SamErrorMetric;
 
 import java.util.Objects;
